requirements.txt
pysolr
numpy
segtok
tqdm
plac
elasticsearch
pandas
bert-serving-server
bert-serving-client
aioelasticsearch
protobuf
